requirements.txt
pandas
numpy
requests
scikit-learn
joblib
vaderSentiment
pynacl
python-dotenv
